Move deprecated generators
